requirements.txt
allure-pytest==2.9.41
allure-python-commons==2.9.41
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-rerunfailures==9.1.1
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
toml==0.10.2
urllib3==1.26.4
